fix: separate upload middleware from validator
